refactor: extract Scorecard components from hole screen (issue #39)
