+++ apps/mobile/app/(app)/round/[id]/hole/[number].tsx
@@ -25,6 +25,10 @@ import {
   PuttingSheet,
   type PuttingValue,
 } from '../../../../../components/round/PuttingSheet'
+import {
+  ScorecardModal,
+  ScorecardPreview,
+} from '../../../../../components/round/Scorecard'
 import { supabase } from '../../../../../lib/supabase'
 import { useAuth } from '../../../../../hooks/useAuth'
 import {
@@ -1141,297 +1145,3 @@ export default function HoleScreen() {
   )
 }
 
-function ScorecardModal({
-  holes,
-  holeScores,
-  currentHoleNumber,
-  onJumpToHole,
-  onClose,
-}: {
-  holes: HoleRow[]
-  holeScores: HoleScoreRow[]
-  currentHoleNumber: number
-  onJumpToHole: (n: number) => void
-  onClose: () => void
-}) {
-  const scoresByHoleId = useMemo(
-    () => new Map(holeScores.map((hs) => [hs.hole_id, hs])),
-    [holeScores],
-  )
-  const sorted = useMemo(
-    () => [...holes].sort((a, b) => a.number - b.number),
-    [holes],
-  )
-  let runningTotal = 0
-  let runningPar = 0
-  return (
-    <View style={{ flex: 1, backgroundColor: 'rgba(0,0,0,0.5)' }}>
-      <Pressable style={{ flex: 1 }} onPress={onClose} />
-      <View
-        style={{
-          backgroundColor: '#FBF8F1',
-          borderTopLeftRadius: 12,
-          borderTopRightRadius: 12,
-          paddingHorizontal: 18,
-          paddingTop: 14,
-          paddingBottom: 28,
-          maxHeight: '85%',
-        }}
-      >
-        <View
-          style={{
-            alignSelf: 'center',
-            width: 32,
-            height: 4,
-            borderRadius: 2,
-            backgroundColor: '#D9D2BF',
-            marginBottom: 14,
-          }}
-        />
-        <Text
-          style={{
-            ...KICKER,
-            color: '#8A8B7E',
-            marginBottom: 6,
-          }}
-        >
-          Scorecard
-        </Text>
-        <ScrollView
-          style={{ maxHeight: '90%' }}
-          showsVerticalScrollIndicator={false}
-        >
-          <View
-            style={{
-              flexDirection: 'row',
-              paddingVertical: 8,
-              borderBottomWidth: 1,
-              borderColor: '#D9D2BF',
-            }}
-          >
-            <Text style={{ ...KICKER, flex: 1, color: '#8A8B7E' }}>Hole</Text>
-            <Text
-              style={{ ...KICKER, width: 44, textAlign: 'right', color: '#8A8B7E' }}
-            >
-              Par
-            </Text>
-            <Text
-              style={{ ...KICKER, width: 56, textAlign: 'right', color: '#8A8B7E' }}
-            >
-              Score
-            </Text>
-            <Text
-              style={{ ...KICKER, width: 56, textAlign: 'right', color: '#8A8B7E' }}
-            >
-              +/−
-            </Text>
-          </View>
-          {sorted.map((h) => {
-            const hs = scoresByHoleId.get(h.id)
-            const score = hs?.score ?? null
-            if (score != null) {
-              runningTotal += score
-              runningPar += h.par
-            }
-            const diff = score != null ? score - h.par : null
-            const active = h.number === currentHoleNumber
-            return (
-              <Pressable
-                key={h.id}
-                accessibilityRole="button"
-                accessibilityLabel={`Jump to hole ${h.number}, par ${h.par}${score != null ? `, score ${score}` : ''}`}
-                accessibilityState={{ selected: active }}
-                onPress={() => onJumpToHole(h.number)}
-                style={{
-                  flexDirection: 'row',
-                  paddingVertical: 10,
-                  borderBottomWidth: 1,
-                  borderColor: '#EBE5D6',
-                  backgroundColor: active ? '#EBE5D6' : 'transparent',
-                  paddingHorizontal: 6,
-                  borderRadius: 2,
-                }}
-              >
-                <Text
-                  style={{
-                    flex: 1,
-                    fontSize: 15,
-                    color: '#1C211C',
-                    fontWeight: active ? '600' : '400',
-                  }}
-                >
-                  {h.number}
-                </Text>
-                <Text
-                  style={{
-                    width: 44,
-                    textAlign: 'right',
-                    fontSize: 15,
-                    color: '#5C6356',
-                    fontVariant: ['tabular-nums'],
-                  }}
-                >
-                  {h.par}
-                </Text>
-                <Text
-                  style={{
-                    width: 56,
-                    textAlign: 'right',
-                    fontSize: 15,
-                    color: score != null ? '#1C211C' : '#8A8B7E',
-                    fontVariant: ['tabular-nums'],
-                    fontWeight: '500',
-                  }}
-                >
-                  {score ?? '—'}
-                </Text>
-                <Text
-                  style={{
-                    width: 56,
-                    textAlign: 'right',
-                    fontSize: 15,
-                    color:
-                      diff == null
-                        ? '#8A8B7E'
-                        : diff < 0
-                          ? '#1F3D2C'
-                          : diff > 0
-                            ? '#A33A2A'
-                            : '#5C6356',
-                    fontVariant: ['tabular-nums'],
-                  }}
-                >
-                  {diff == null ? '—' : diff === 0 ? 'E' : diff > 0 ? `+${diff}` : `${diff}`}
-                </Text>
-              </Pressable>
-            )
-          })}
-          <View
-            style={{
-              flexDirection: 'row',
-              paddingVertical: 12,
-              borderTopWidth: 1,
-              borderColor: '#9F9580',
-              marginTop: 4,
-              paddingHorizontal: 6,
-            }}
-          >
-            <Text style={{ flex: 1, fontSize: 14, fontWeight: '600', color: '#1C211C' }}>
-              Total played
-            </Text>
-            <Text
-              style={{
-                width: 44,
-                textAlign: 'right',
-                fontSize: 14,
-                fontWeight: '600',
-                color: '#1C211C',
-                fontVariant: ['tabular-nums'],
-              }}
-            >
-              {runningPar}
-            </Text>
-            <Text
-              style={{
-                width: 56,
-                textAlign: 'right',
-                fontSize: 14,
-                fontWeight: '600',
-                color: '#1C211C',
-                fontVariant: ['tabular-nums'],
-              }}
-            >
-              {runningTotal}
-            </Text>
-            <Text
-              style={{
-                width: 56,
-                textAlign: 'right',
-                fontSize: 14,
-                fontWeight: '600',
-                color:
-                  runningPar === 0
-                    ? '#8A8B7E'
-                    : runningTotal - runningPar < 0
-                      ? '#1F3D2C'
-                      : runningTotal - runningPar > 0
-                        ? '#A33A2A'
-                        : '#5C6356',
-                fontVariant: ['tabular-nums'],
-              }}
-            >
-              {runningPar === 0
-                ? '—'
-                : runningTotal === runningPar
-                  ? 'E'
-                  : runningTotal - runningPar > 0
-                    ? `+${runningTotal - runningPar}`
-                    : `${runningTotal - runningPar}`}
-            </Text>
-          </View>
-        </ScrollView>
-        <Pressable
-          accessibilityRole="button"
-          accessibilityLabel="Close scorecard"
-          onPress={onClose}
-          style={{
-            marginTop: 14,
-            paddingVertical: 12,
-            alignItems: 'center',
-            borderWidth: 1,
-            borderColor: '#D9D2BF',
-            borderRadius: 2,
-          }}
-        >
-          <Text style={{ ...KICKER, color: '#5C6356' }}>Close</Text>
-        </Pressable>
-      </View>
-    </View>
-  )
-}
-
-function ScorecardPreview({
-  holes,
-  holeScores,
-  currentHoleNumber,
-}: {
-  holes: HoleRow[]
-  holeScores: HoleScoreRow[]
-  currentHoleNumber: number
-}) {
-  const scoresByHoleId = new Map(holeScores.map((hs) => [hs.hole_id, hs]))
-  return (
-    <ScrollView horizontal showsHorizontalScrollIndicator={false}>
-      <View style={{ flexDirection: 'row', alignItems: 'center', gap: 4 }}>
-        {holes.map((h) => {
-          const hs = scoresByHoleId.get(h.id)
-          const active = h.number === currentHoleNumber
-          return (
-            <View
-              key={h.id}
-              style={{
-                width: 26,
-                height: 26,
-                alignItems: 'center',
-                justifyContent: 'center',
-                borderRadius: 2,
-                backgroundColor: active ? '#1F3D2C' : '#EBE5D6',
-              }}
-            >
-              <Text
-                style={{
-                  fontSize: 11,
-                  fontWeight: '500',
-                  color: active ? '#F2EEE5' : '#5C6356',
-                  fontVariant: ['tabular-nums'],
-                }}
-              >
-                {hs && hs.score ? hs.score : h.number}
-              </Text>
-            </View>
-          )
-        })}
-      </View>
-    </ScrollView>
-  )
-}
